requirement.txt
matplotlib==3.4.3
pandas==1.3.0
reportlab==3.5.67
Pillow==8.2.0
